+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Added ability to import students to an organization from `.csv` files
+
 ## 0.37.0 - Student Enrollments Overhaul
 - Changed serializing and scoping for students to go through organization and by groups found in their enrollments
 - Modified generation of new MLIDs to go through `organization_id` for students