+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+- Added dropdown to select enrollment starting date for new students in a group
+- Modified CI workflow with caching and cleanup separation, added retries using `rspec-retry` gem
 - Added ability to enroll students with no active enrollments in a group
 - Added ability to import students to an organization from `.csv` files
 - Added organization statistics in separate component when viewing an organization